Flesh out elf generation
Wood elves and high elves have different usual characteristics; so far
it has been left such that either one can have a common skin type, and a
common hair colour with it. This means there will be some
out-of-character elves generated (for instance, a white wood-elf with
blue hair) but such quirks are to be expected sometimes in nature (and
fixing that sort of thing would be too fiddly for the amount of value it
would provide).